update water yield to ZLD
